refactor: Address mapper-seam review findings
Relocate buildPermissionApprovalResponse (and its private text helper) from
the tools layer to utils/payment_errors.ts so the pure error mapper no longer
transitively pulls in call_actor's graph; move its unit suite next to the
payment builder's. Cover the mapper's execution failure_http_status arm, hoist
the shared X402_PAYMENT_DATA fixture into the test helper, add the
TOOL_CALL_ERROR_KIND type alias, and correct the AGENTS.md wire-wrapping note,
the mapper JSDoc, the server.ts hoist comment, and makePaymentRequiredError's
no-arg contract.

ACTOR_MCP telemetry unification, dispatch exhaustiveness, and a handler-level
containment regression test are deferred to a follow-up issue.
